+++ CHANGELOG.md
@@ -16,6 +16,10 @@
   `--swift-build-tool symbolgraph`.  
   [John Fairhurst](https://github.com/johnfairh)
 
+* Support Swift `@_documentation(visibility:)` attribute.  Requires
+  `--swift-build-tool spm|xcodebuild`.  
+  [John Fairhurst](https://github.com/johnfairh)
+
 ##### Bug Fixes
 
 * Issue a warning instead of crashing on declarations without names.